Move components into folders, remove page components
