+++ src/js/edu-benefits/components/military-history/DependentInformationFields.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import ErrorableRadioButtons from '../../common/components/form-elements/ErrorableRadioButtons';
+import ErrorableRadioButtons from '../../../common/components/form-elements/ErrorableRadioButtons';
 
-import { yesNo } from '../utils/options-for-select';
+import { yesNo } from '../../utils/options-for-select';
 
 export default class DependentInformationFields extends React.Component {
   render() {
